Actually use geolocation on the map
